Descount in process on the development

+++ lib/data/geral.data.dart
@@ -1,16 +1,16 @@
 class GeralData {
   List<dynamic> _isMaintenance;
-  int _discountGeral;
+  double _discountGeral;
 
-  GeralData({List<dynamic> isMaintenance, int discountGeral}) {
+  GeralData({List<dynamic> isMaintenance, double discountGeral}) {
     this._isMaintenance = isMaintenance;
     this._discountGeral = discountGeral;
   }
 
   List<dynamic> get isMaintenance => _isMaintenance;
   set isMaintenance(List<dynamic> isMaintenance) => _isMaintenance = isMaintenance;
-  int get discountGeral => _discountGeral;
-  set discountGeral(int discountGeral) => _discountGeral = discountGeral;
+  double get discountGeral => _discountGeral;
+  set discountGeral(double discountGeral) => _discountGeral = discountGeral;
 
   GeralData.fromJson(Map<String, dynamic> json) {
     _isMaintenance = json['isMaintenance'].cast<dynamic>();
